+++ apps/playground-web/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @motif-js/playground-web
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @usemotif/headless@1.1.1
+  - @usemotif/tokens@1.1.1
+  - @usemotif/icons@1.1.1
+  - usemotif@1.1.1
+
 ## 0.0.5
 
 ### Patch Changes