+++ apps/ssr-next/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @motif-js/ssr-next
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @usemotif/react@1.1.1
+  - @usemotif/tokens@1.1.1
+
 ## 0.0.5
 
 ### Patch Changes